Restructure relay module
