+++ src/relay/index.ts
@@ -1,0 +1,1 @@
+export { applyCursorsToEdges, relayPaginateNodes } from './utils';
