+++ src/relay/utils.ts
@@ -1,14 +1,4 @@
-export type Edge<T> = { node: T; cursor: string };
-export type CursorForNode<T> = (node: T) => string;
-export type RelayPaginationResult<T = unknown> = {
-  edges: { node: T }[];
-  pageInfo: {
-    hasNextPage: boolean;
-    hasPreviousPage: boolean;
-    startCursor: string | null;
-    endCursor: string | null;
-  };
-};
+import { Edge, CursorForNode, RelayPaginationResult } from './types';
 
 export const createEdge = <T>(node: T, cursor: string): { cursor: string; node: T } => ({
   cursor,
